web: update navbar header into sub-components for ease of styling
This commit breaks ak-page-navbar up into seperate units, mapping to the parts of an HTMLElement:
The Style block, the HTML template, and the business logic.  Doing this makes a *whole lot* of
compromises and hacks stand out, like the whaty we backfill uiConfig.navbar.userdisplay, or
the way icons are just a mess.

No business logic was revised with this commit *at all*.  This is simply breaking the component
up so that the logic is exposed.

\# What

\# Why

\# How

\# Designs

\# Test Steps

\# Other Notes