ci: fix checkout stable
If you don't happen to read the logs of this step, everything seems
fine and `git checkout $(prev_stable)` is just interpreted as the
no-op `git checkout`. That's because `$()` spawns a subshell and
subshell errors don't stop execution on `set -eo pipefail`. In fact, a
quick search didn't reveal any way to catch this typo with `set`.

I wonder how long this could have stayed hidden if I wasn't specifically
looking at this workflow for a completely different reason.