admin: make empty table message configurable
This commit provides a new field at the Table level for the empty state message. The field defaults to the original message, “No objects found.”

The icon has long been configurable, but not the message. It makes sense to customize this message and let people know if they’re looking at files, properties, applications, and other objects.

Co-authored-by: Ken Sternberg <ken@goauthentik.io>